Move ffi to utils
