--- py-polars/polars/ffi.py
@@ -1,28 +1,0 @@
-import ctypes
-from typing import Any
-
-import numpy as np
-
-# https://stackoverflow.com/questions/4355524/getting-data-from-ctypes-array-into-numpy
-
-
-def _ptr_to_numpy(ptr: int, len: int, ptr_type: Any) -> np.ndarray:
-    """
-
-    Parameters
-    ----------
-    ptr
-        C/Rust ptr casted to usize.
-    len
-        Length of the array values.
-    ptr_type
-        Example:
-            f32: ctypes.c_float)
-
-    Returns
-    -------
-    View of memory block as numpy array.
-
-    """
-    ptr_ctype = ctypes.cast(ptr, ctypes.POINTER(ptr_type))
-    return np.ctypeslib.as_array(ptr_ctype, (len,))
